first try to create a validator
