+++ clinicadl/trainer/trainer.py
@@ -33,6 +33,7 @@ from clinicadl.utils.iotools.trainer_utils import (
     patch_to_read_json,
 )
 from clinicadl.trainer.tasks_utils import create_training_config
+from clinicadl.validator.validator import Validator
 
 if TYPE_CHECKING:
     from clinicadl.callbacks.callbacks import Callback
@@ -64,6 +65,7 @@ class Trainer:
         """
         self.config = config
         self.maps_manager = self._init_maps_manager(config)
+        self.validator = Validator()
         self._check_args()
 
     def _init_maps_manager(self, config) -> MapsManager:
@@ -371,12 +373,14 @@ class Trainer:
             )
 
             if cluster.master:
-                self.maps_manager._ensemble_prediction(
+                self.validator._ensemble_prediction(
+                    self.maps_manager,
                     "train",
                     split,
                     self.config.validation.selection_metrics,
                 )
-                self.maps_manager._ensemble_prediction(
+                self.validator._ensemble_prediction(
+                    self.maps_manager,
                     "validation",
                     split,
                     self.config.validation.selection_metrics,
@@ -495,12 +499,14 @@ class Trainer:
                 resume = False
 
             if cluster.master:
-                self.maps_manager._ensemble_prediction(
+                self.validator._ensemble_prediction(
+                    self.maps_manager,
                     "train",
                     split,
                     self.config.validation.selection_metrics,
                 )
-                self.maps_manager._ensemble_prediction(
+                self.validator._ensemble_prediction(
+                    self.maps_manager,
                     "validation",
                     split,
                     self.config.validation.selection_metrics,
@@ -706,12 +712,14 @@ class Trainer:
                 resume=resume,
             )
 
-            self.maps_manager._ensemble_prediction(
+            self.validator._ensemble_prediction(
+                self.maps_manager,
                 "train",
                 split,
                 self.config.validation.selection_metrics,
             )
-            self.maps_manager._ensemble_prediction(
+            self.validator._ensemble_prediction(
+                self.maps_manager,
                 "validation",
                 split,
                 self.config.validation.selection_metrics,
@@ -998,7 +1006,8 @@ class Trainer:
             epoch += 1
 
         del model
-        self.maps_manager._test_loader(
+        self.validator._test_loader(
+            self.maps_manager,
             train_loader,
             criterion,
             "train",
@@ -1007,7 +1016,8 @@ class Trainer:
             amp=self.maps_manager.std_amp,
             network=network,
         )
-        self.maps_manager._test_loader(
+        self.validator._test_loader(
+            self.maps_manager,
             valid_loader,
             criterion,
             "validation",
@@ -1018,7 +1028,8 @@ class Trainer:
         )
 
         if save_outputs(self.maps_manager.network_task):
-            self.maps_manager._compute_output_tensors(
+            self.validator._compute_output_tensors(
+                self.maps_manager,
                 train_loader.dataset,
                 "train",
                 split,
@@ -1026,7 +1037,8 @@ class Trainer:
                 nb_images=1,
                 network=network,
             )
-            self.maps_manager._compute_output_tensors(
+            self.validator._compute_output_tensors(
+                self.maps_manager,
                 valid_loader.dataset,
                 "validation",
                 split,
@@ -1400,7 +1412,8 @@ class Trainer:
 
             epoch += 1
 
-        self.maps_manager._test_loader_ssda(
+        self.validator._test_loader_ssda(
+            self.maps_manager,
             train_target_loader,
             criterion,
             data_group="train",
@@ -1410,7 +1423,8 @@ class Trainer:
             target=True,
             alpha=0,
         )
-        self.maps_manager._test_loader_ssda(
+        self.validator._test_loader_ssda(
+            self.maps_manager,
             valid_loader,
             criterion,
             data_group="validation",
@@ -1422,7 +1436,8 @@ class Trainer:
         )
 
         if save_outputs(self.maps_manager.network_task):
-            self.maps_manager._compute_output_tensors(
+            self.validator._compute_output_tensors(
+                self.maps_manager,
                 train_target_loader.dataset,
                 "train",
                 split,
@@ -1430,7 +1445,8 @@ class Trainer:
                 nb_images=1,
                 network=network,
             )
-            self.maps_manager._compute_output_tensors(
+            self.validator._compute_output_tensors(
+                self.maps_manager,
                 train_target_loader.dataset,
                 "validation",
                 split,
